refactor(calendar): move pager into own package
