+++ calendar/src/main/java/com/jonaswanke/calendar/pager/InfiniteViewPager.kt
@@ -1,4 +1,4 @@
-package com.jonaswanke.calendar
+package com.jonaswanke.calendar.pager
 
 import android.content.Context
 import android.os.Bundle
@@ -8,6 +8,7 @@ import androidx.viewpager.widget.ViewPager
 import android.util.AttributeSet
 import android.util.Log
 import android.view.View
+import com.jonaswanke.calendar.BuildConfig
 import kotlinx.coroutines.experimental.android.UI
 import kotlinx.coroutines.experimental.launch
