+++ calendar/src/main/java/com/jonaswanke/calendar/pager/PageModel.kt
@@ -1,4 +1,4 @@
-package com.jonaswanke.calendar
+package com.jonaswanke.calendar.pager
 
 
 import android.view.View
